refactor symbols logic
