+++ lua/codecompanion/strategies/chat/slash_commands/help.lua
@@ -1,4 +1,4 @@
-local path = require("plenary.path")
+local Path = require("plenary.path")
 
 local config = require("codecompanion.config")
 local log = require("codecompanion.utils.log")
@@ -106,7 +106,7 @@ local function output(SlashCommand, selected)
     return log:warn("Sending of code has been disabled")
   end
 
-  local content = path.new(selected.path):read()
+  local content = Path.new(selected.path):read()
   line_count = #vim.split(content, "\n")
 
   if line_count > CONSTANTS.MAX_LINES then
